Move `OrchestrationResult` to `schemas.responses` to fix circular imports
