+++ src/prefect/client/schemas.py
@@ -5,9 +5,6 @@ from typing import Generic, Iterable, Optional, Type, TypeVar, Union, overload
 from pydantic import Field
 
 from prefect.orion import schemas
-from prefect.orion.orchestration.rules import (
-    OrchestrationResult as ServerOrchestrationResult,
-)
 
 R = TypeVar("R")
 
@@ -232,5 +229,5 @@ class TaskRun(schemas.core.TaskRun.subclass()):
     state: Optional[State] = Field(default=None)
 
 
-class OrchestrationResult(ServerOrchestrationResult.subclass()):
+class OrchestrationResult(schemas.responses.OrchestrationResult.subclass()):
     state: Optional[State]
